+++ backend/app/src/test/java/coffeeshout/blindtimer/ui/BlindTimerGameIntegrationTest.java
@@ -1,13 +1,13 @@
-package coffeeshout.blindtimer.ui;
+﻿package coffeeshout.blindtimer.ui;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
 import coffeeshout.blindtimer.domain.BlindTimerGame;
 import coffeeshout.fixture.RoomFixture;
-import coffeeshout.fixture.TestStompSession;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.fixture.WebSocketIntegrationTestSupport;
-import coffeeshout.global.MessageResponse;
+import coffeeshout.support.MessageResponse;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.player.Player;
