+++ backend/app/src/test/java/coffeeshout/blockstacking/ui/BlockStackingIntegrationTest.java
@@ -1,12 +1,12 @@
-package coffeeshout.blockstacking.ui;
+﻿package coffeeshout.blockstacking.ui;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import coffeeshout.blockstacking.domain.BlockStackingGame;
 import coffeeshout.fixture.RoomFixture;
-import coffeeshout.fixture.TestStompSession;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.fixture.WebSocketIntegrationTestSupport;
-import coffeeshout.global.MessageResponse;
+import coffeeshout.support.MessageResponse;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.player.Player;
