+++ backend/app/src/test/java/coffeeshout/fixture/WebSocketIntegrationTestSupport.java
@@ -1,8 +1,9 @@
-package coffeeshout.fixture;
+﻿package coffeeshout.fixture;
 
 import static org.skyscreamer.jsonassert.JSONCompareMode.LENIENT;
 
-import coffeeshout.global.MessageResponse;
+import coffeeshout.support.MessageResponse;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.room.infra.auth.RoomSessionTokenService;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.player.PlayerName;
