+++ backend/app/src/test/java/coffeeshout/racinggame/ui/RacingGameIntegrationTest.java
@@ -1,9 +1,9 @@
-package coffeeshout.racinggame.ui;
+﻿package coffeeshout.racinggame.ui;
 
 import coffeeshout.fixture.RoomFixture;
-import coffeeshout.fixture.TestStompSession;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.fixture.WebSocketIntegrationTestSupport;
-import coffeeshout.global.MessageResponse;
+import coffeeshout.support.MessageResponse;
 import coffeeshout.racinggame.domain.RacingGame;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.Room;
