+++ backend/app/src/test/java/coffeeshout/room/ui/RoomWebSocketControllerTest.java
@@ -2,8 +2,8 @@ package coffeeshout.room.ui;
 
 import coffeeshout.fixture.RoomFixture;
 import coffeeshout.fixture.WebSocketIntegrationTestSupport;
-import coffeeshout.global.MessageResponse;
-import coffeeshout.fixture.TestStompSession;
+import coffeeshout.support.MessageResponse;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.RoomState;
