+++ backend/app/src/test/java/coffeeshout/speedtouch/ui/SpeedTouchGameIntegrationTest.java
@@ -1,12 +1,12 @@
-package coffeeshout.speedtouch.ui;
+﻿package coffeeshout.speedtouch.ui;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
 import coffeeshout.fixture.RoomFixture;
-import coffeeshout.fixture.TestStompSession;
+import coffeeshout.support.TestStompSession;
 import coffeeshout.fixture.WebSocketIntegrationTestSupport;
-import coffeeshout.global.MessageResponse;
+import coffeeshout.support.MessageResponse;
 import coffeeshout.room.domain.JoinCode;
 import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.player.Player;
